move FindCertificateByThumbprint to CertificateUtility
