+++ src/NuGet.Services.KeyVault/NuGet.Services.KeyVault.csproj
@@ -43,6 +43,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CachingSecretReader.cs" />
+    <Compile Include="CertificateUtility.cs" />
     <Compile Include="ISecretReaderFactory.cs" />
     <Compile Include="EmptySecretReader.cs" />
     <Compile Include="ISecretInjector.cs" />
